Stub in @aws validation
