+++ src/config/pragmas/validate/index.js
@@ -1,22 +1,12 @@
-let patterns = {
-  looseName: new RegExp(/^[a-z][a-zA-Z0-9-_]+$/),
-  strictName: new RegExp(/^[a-z][a-z0-9-]+$/),
-}
-
-function regex (value, pattern, pragma) {
-  let matching = typeof pattern === 'string' ? patterns[pattern] : pattern
-  if (!matching.test(value)) {
-    throw Error(`Invalid ${pragma} value: ${value} must match ${pattern}`)
-  }
-}
-
-function size (value, num, pragma) {
-  if (typeof value !== 'string') throw Error(`Value must be a string: ${value}`)
-  if (typeof num !== 'number') throw Error(`validate.size requires a number: ${num}`)
-  if (value.length > num) throw Error(`Invalid ${pragma} value: ${value} must less than ${num} characters`)
-}
+let aws = require('./_aws')
+let { patterns, regex, size } = require('./_meta')
 
 module.exports = {
+  // Pragmas and project validation
+  aws,
+
+  // TODO prime refactor candidate as we get deeper into validation:
+  // Meta
   patterns,
   validate: {
     regex,
